factor icons and markers into own files
