+++ vite.config.js
@@ -7,7 +7,6 @@ export default defineConfig({
   build: {
     outDir: '../dist',
     assetsDir: 'assets',
-    manifest: true,
   },
   plugins: [replaceGeoJsonInHtml()],
